+++ Deeploy/MemoryLevelExtension/MemoryLevels.py
@@ -2,12 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from typing import Dict, List, Optional, Sequence, Tuple
-
-import onnx_graphsurgeon as gs
-
-from Deeploy.DeeployTypes import CodeTransformation, NetworkContext, NodeBinding, NodeTemplate, NodeTypeChecker, \
-    OperatorRepresentation
+from typing import Dict, List, Optional
 
 
 class MemoryLevel():
@@ -109,58 +104,3 @@ class MemoryHierarchy():
         if self._defaultMemoryLevel is None:
             raise ValueError('defaultMemoryLevel level not set!')
         return self._defaultMemoryLevel
-
-
-class NodeMemoryLevelChecker():
-
-    def __init__(self, inputMemoryLevels: Sequence[Optional[str]], outputMemoryLevels: Sequence[Optional[str]]):
-        self.inputMemoryLevels = inputMemoryLevels
-        self.outputMemoryLevels = outputMemoryLevels
-
-    def _memEq(self, memoryLevel: str, annotatedMemoryLevel: str) -> bool:
-        if memoryLevel is None:
-            return True
-        else:
-            return memoryLevel == annotatedMemoryLevel
-
-    def _checkMemoryLevels(self, ctxt: NetworkContext, memoryLevels: Sequence[str],
-                           tensors: Sequence[gs.Tensor]) -> bool:
-        buffers = [ctxt.lookup(tensor.name) for tensor in tensors]
-        if not all(hasattr(buffer, "_memoryLevel") for buffer in buffers):
-            return False
-
-        annotatedMemoryLevels = [buffer._memoryLevel for buffer in buffers]
-        if all(
-                self._memEq(memoryLevel, annotatedMemoryLevel)
-                for memoryLevel, annotatedMemoryLevel in zip(memoryLevels, annotatedMemoryLevels)):
-            return True
-        else:
-            return False
-
-    def check(self, ctxt: NetworkContext, node: gs.Node, operatorRepresentation) -> Tuple[NetworkContext, bool]:
-        if self._checkMemoryLevels(ctxt, self.inputMemoryLevels, node.inputs) and self._checkMemoryLevels(
-                ctxt, self.outputMemoryLevels, node.outputs):
-            return ctxt, True
-        else:
-            return ctxt, False
-
-
-class MemoryAwareNodeBinding(NodeBinding):
-
-    def __init__(self, typeChecker: NodeTypeChecker, memoryLevelChecker: NodeMemoryLevelChecker, template: NodeTemplate,
-                 codeTransformer: CodeTransformation):
-        super().__init__(typeChecker, template, codeTransformer)
-        self.memoryLevelChecker = memoryLevelChecker
-
-    def typeCheck(self, ctxt: NetworkContext, node: gs.Node,
-                  operatorRepresentation: OperatorRepresentation) -> Tuple[NetworkContext, bool]:
-        newCtxt, ret = self.memoryLevelChecker.check(ctxt, node, operatorRepresentation)
-        if ret:
-            return super().typeCheck(newCtxt, node, operatorRepresentation)
-
-        return ctxt, False
-
-
-def memoryAwareNodeBindingExtension(binding: NodeBinding,
-                                    memoryLevelChecker: NodeMemoryLevelChecker) -> MemoryAwareNodeBinding:
-    return MemoryAwareNodeBinding(binding.typeChecker, memoryLevelChecker, binding.template, binding.codeTransformer)