+++ Deeploy/MemoryLevelExtension/NetworkDeployers/MemoryLevelDeployer.py
@@ -11,7 +11,7 @@ from Deeploy.AbstractDataTypes import Pointer
 from Deeploy.CommonExtensions.NetworkDeployers.NetworkDeployerWrapper import NetworkDeployerWrapper
 from Deeploy.CommonExtensions.NetworkDeployers.SignPropDeployer import SignPropDeployer
 from Deeploy.DeeployTypes import CodeGenVerbosity, ConstantBuffer, DeploymentEngine, DeploymentPlatform, \
-    NetworkContext, NetworkDeployer, NetworkOptimizationPass, NetworkOptimizer, ONNXLayer, Schedule, StructBuffer, \
+    NetworkContext, NetworkDeployer, NetworkOptimizationPass, NetworkOptimizer, Schedule, StructBuffer, \
     TopologyOptimizer, TransientBuffer, VariableBuffer, _NoVerbosity
 from Deeploy.Logging import DEFAULT_LOGGER as log
 from Deeploy.MemoryLevelExtension.MemoryLevels import MemoryHierarchy, MemoryLevel
@@ -128,25 +128,16 @@ class MemoryLevelAwareDeployer(NetworkDeployer, MemorySummaryMixin):
             f"Platform should be a MemoryPlatform or MemoryPlatformWrapper! Got {type(self.Platform).__name__}"
         return TargetMemoryLevelMapping(self.graph, self.Platform, self.ctxt)
 
-    def _parseNode(self, node: ONNXLayer, ctxt: NetworkContext,
-                   default_channels_first: bool) -> Tuple[NetworkContext, bool]:
-
-        newCtxt, parsePass = super()._parseNode(node, ctxt, default_channels_first)
-
-        if not parsePass:
-            return ctxt, False
-
-        newCtxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(newCtxt, self.graph)
-
-        return newCtxt, parsePass
-
     def bind(self):
+        log.info("- Perform Memory Level Annotation")
+        # LMACAN: Annotate before bind because during binding (specifically alignToContext) templates
+        #         may expect the memoryLevel annotation already.
+        self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)
 
         ret = super().bind()
         if not ret:
             return False
 
-        log.info("- Perform Memory Level Annotation")
         # SCHEREMO: There might be hoisting; reassign memoryLevel preferences
         self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)
 
@@ -181,29 +172,16 @@ class MemoryLevelAwareSignPropDeployer(SignPropDeployer, MemorySummaryMixin):
             f"Platform should be a MemoryPlatform or MemoryPlatformWrapper! Got {type(self.Platform).__name__}"
         return TargetMemoryLevelMapping(self.graph, self.Platform, self.ctxt)
 
-    def _parseNode(self, node: ONNXLayer, ctxt: NetworkContext,
-                   default_channels_first: bool) -> Tuple[NetworkContext, bool]:
-
-        newCtxt, parsePass = node.parse(ctxt.copy(), default_channels_first)
-
-        if not parsePass:
-            return ctxt, False
-
-        newCtxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(newCtxt, self.graph)
-        newCtxt, LayerBindSuccess = node.typeCheck(newCtxt)
-
-        if not LayerBindSuccess:
-            return ctxt, False
-
-        return newCtxt, True
-
     def bind(self):
+        log.info("- Perform Memory Level Annotation")
+        # LMACAN: Annotate before bind because during binding (specifically alignToContext) templates
+        #         may expect the memoryLevel annotation already.
+        self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)
 
         ret = super().bind()
         if not ret:
             return False
 
-        log.info("- Perform Memory Level Annotation")
         # SCHEREMO: There might be hoisting; reassign memoryLevel preferences
         self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)
 
@@ -229,29 +207,16 @@ class MemoryDeployerWrapper(NetworkDeployerWrapper, MemorySummaryMixin):
             f"Platform should be a MemoryPlatform or MemoryPlatformWrapper! Got {type(self.Platform).__name__}"
         return TargetMemoryLevelMapping(self.graph, self.Platform, self.ctxt)
 
-    def _parseNode(self, node: ONNXLayer, ctxt: NetworkContext,
-                   default_channels_first: bool) -> Tuple[NetworkContext, bool]:
-
-        newCtxt, parsePass = node.parse(ctxt.copy(), default_channels_first)
-
-        if not parsePass:
-            return ctxt, False
-
-        newCtxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(newCtxt, self.graph)
-        newCtxt, LayerBindSuccess = node.typeCheck(newCtxt)
-
-        if not LayerBindSuccess:
-            return ctxt, False
-
-        return newCtxt, True
-
     def bind(self):
+        log.info("- Perform Memory Level Annotation")
+        # LMACAN: Annotate before bind because during binding (specifically alignToContext) templates
+        #         may expect the memoryLevel annotation already.
+        self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)
 
         ret = super().bind()
         if not ret:
             return False
 
-        log.info("- Perform Memory Level Annotation")
         # SCHEREMO: There might be hoisting; reassign memoryLevel preferences
         self.ctxt, self.graph = self.memoryLevelAnnotationOptimizer.optimize(self.ctxt, self.graph)